feat: loading neurons
