+++ frontend/src/lib/components/common/SignedInOnly.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-    import { isSignedIn } from "$lib/utils/auth.utils";
-    import { authStore } from "$lib/stores/auth.store";
+  import { isSignedIn } from "$lib/utils/auth.utils";
+  import { authStore } from "$lib/stores/auth.store";
 
-    let signedIn = false;
-    $: signedIn = isSignedIn($authStore.identity);
+  let signedIn = false;
+  $: signedIn = isSignedIn($authStore.identity);
 </script>
 
 {#if signedIn}
-    <slot />
+  <slot />
 {/if}
