+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingCard.svelte
@@ -6,7 +6,7 @@
   } from "@dfinity/nns";
 
   import { getContext, onDestroy } from "svelte";
-  import { definedNeuronsStore } from "$lib/stores/neurons.store";
+  import { definedNeuronsStore, neuronsStore } from "$lib/stores/neurons.store";
   import { votingNeuronSelectStore } from "$lib/stores/proposals.store";
   import VotingConfirmationToolbar from "./VotingConfirmationToolbar.svelte";
   import VotingNeuronSelect from "./VotingNeuronSelect.svelte";
@@ -21,6 +21,8 @@
   } from "$lib/stores/vote-registration.store";
   import { registerVotes } from "$lib/services/vote-registration.services";
   import { BottomSheet } from "@dfinity/gix-components";
+  import { Spinner } from "@dfinity/gix-components";
+  import { i18n } from "$lib/stores/i18n";
 
   export let proposalInfo: ProposalInfo;
 
@@ -76,18 +78,59 @@
     unsubscribe();
     votingNeuronSelectStore.reset();
   });
+
+  // UI loader
+  const neuronsStoreReady = (): boolean => {
+    // We consider the neurons store as ready if it has been initialized once. Subsequent changes that happen after vote or other functions are handled with the busy store.
+    // This to avoid the display of a spinner within the page and another spinner over it (the busy spinner) when the user vote is being processed.
+    if (neuronsReady) {
+      return true;
+    }
+
+    return (
+      $neuronsStore.neurons !== undefined && $neuronsStore.certified === true
+    );
+  };
+
+  let neuronsReady = false;
+  $: $neuronsStore, (neuronsReady = neuronsStoreReady());
 </script>
 
 {#if $definedNeuronsStore.length > 0}
   <BottomSheet>
-    {#if visible}
-      <VotingConfirmationToolbar
-        {proposalInfo}
-        {voteRegistration}
-        on:nnsConfirm={vote}
-      />
-    {/if}
+    {#if neuronsReady}
+      {#if visible}
+        <VotingConfirmationToolbar
+          {proposalInfo}
+          {voteRegistration}
+          on:nnsConfirm={vote}
+        />
+      {/if}
 
-    <VotingNeuronSelect {proposalInfo} {voteRegistration} />
+      <VotingNeuronSelect {proposalInfo} {voteRegistration} />
+    {:else}
+      <div class="loader">
+        <span class="spinner"><Spinner inline size="small" /></span>
+        <span> <small>{$i18n.proposal_detail.loading_neurons}</small></span>
+      </div>
+    {/if}
   </BottomSheet>
 {/if}
+
+<style lang="scss">
+  @use "@dfinity/gix-components/styles/mixins/media";
+
+  .loader {
+    padding: calc(4.25 * var(--padding)) var(--padding-2x) var(--padding);
+
+    @include media.min-width(large) {
+      padding: var(--padding-3x) 0;
+    }
+  }
+
+  .spinner {
+    display: inline-block;
+    position: relative;
+    vertical-align: middle;
+  }
+</style>
