+++ frontend/src/lib/components/proposals/ProposalsFilters.svelte
@@ -82,10 +82,10 @@
 
   <SignedInOnly>
     <Checkbox
-            inputId="hide-unavailable-proposals"
-            checked={excludeVotedProposals}
-            on:nnsChange={() => proposalsFiltersStore.toggleExcludeVotedProposals()}
-            text="block">{$i18n.voting.hide_unavailable_proposals}</Checkbox
+      inputId="hide-unavailable-proposals"
+      checked={excludeVotedProposals}
+      on:nnsChange={() => proposalsFiltersStore.toggleExcludeVotedProposals()}
+      text="block">{$i18n.voting.hide_unavailable_proposals}</Checkbox
     >
   </SignedInOnly>
 </div>
