+++ frontend/src/routes/(app)/proposal/+page.svelte
@@ -1,15 +1,9 @@
 <script lang="ts">
-  import SignInProposals from "$lib/pages/SignInProposals.svelte";
-  import { isSignedIn } from "$lib/utils/auth.utils";
-  import { authStore } from "$lib/stores/auth.store";
   import RouteModule from "$lib/components/common/RouteModule.svelte";
   import { AppPath } from "$lib/constants/routes.constants";
   import { afterNavigate } from "$app/navigation";
   import type { Navigation } from "@sveltejs/kit";
   import { referrerPathForNav } from "$lib/utils/page.utils";
-
-  let signedIn = false;
-  $: signedIn = isSignedIn($authStore.identity);
 
   // Preloaded by +page.ts
   export let data: { proposal: string | null | undefined };
@@ -21,11 +15,7 @@
   afterNavigate((nav: Navigation) => (referrerPath = referrerPathForNav(nav)));
 </script>
 
-{#if signedIn}
-  <RouteModule
-    path={AppPath.Proposal}
-    params={{ proposalIdText: proposalId, referrerPath }}
-  />
-{:else}
-  <SignInProposals />
-{/if}
+<RouteModule
+  path={AppPath.Proposal}
+  params={{ proposalIdText: proposalId, referrerPath }}
+/>
